feat: search results pagination
Implement a kind of offset-based pagination for search results. Page
size is 20 but can be adusted later if we want.

Since we can't change the tests, we also can't change the response shape
of the endpoint, which is limited to an array. To pass the offset value
to the client, we rely on a custom header value instead.

The client now holds a couple of things in the memory:

1. The results in memory because they are paged and new results should
   be merged with the previous ones.
2. A boolean flag whether there are more results to fetch.

I decided to separate `loadMore` and `search` because, though they share
_some_ behaviors, they have enough distinctions to where combining them
might result in a harder-to-comprehend code.